+++ project1_GameRoom/Project/backend/조성환/blog/src/main/java/com/cos/blog/dto/BoastBoardPageResponseDto.java
@@ -1,0 +1,16 @@
+package com.cos.blog.dto;
+
+import org.springframework.data.domain.Page;
+
+import com.cos.blog.model.BoastBoard;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class BoastBoardPageResponseDto {
+	Page<BoastBoard> page;
+}
