+++ project1_GameRoom/Project/backend/조성환/blog/src/main/java/com/cos/blog/dto/BoastBoardResponseDto.java
@@ -1,0 +1,14 @@
+package com.cos.blog.dto;
+
+import com.cos.blog.model.BoastBoard;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class BoastBoardResponseDto {
+	BoastBoard boastBoard;
+}
